# Add fast, layer-merged forward and inverse NTT code.