--- a/cbmc/Makefile.common
+++ b/cbmc/Makefile.common
@@ -949,6 +949,7 @@ result:
 	$(MAKE) -B _result
 	@ echo Running 'litani build'
 	$(LITANI) run-build
+	tail -4 $(LOGDIR)/result.txt
 
 _property: $(LOGDIR)/property.xml
 property: