--- a/cbmc/basemul_cached/basemul_cached_harness.c
+++ b/cbmc/basemul_cached/basemul_cached_harness.c
@@ -6,6 +6,9 @@
 #include "ntt.h"
 #include "reduce.h"
 
+void basemul_cached(int16_t r[2], const int16_t a[2], const int16_t b[2],
+                    const int16_t b_cached);
+
 void harness(void)
 {
   int16_t *a, *b, *r, b_cached;