--- a/cbmc/invntt_layer6/cbmc-proof.txt
+++ b/cbmc/invntt_layer6/cbmc-proof.txt
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+# This file marks this directory as containing a CBMC proof.