--- a/cbmc/invntt_layer7_invert/invntt_layer7_invert_harness.c
+++ b/cbmc/invntt_layer7_invert/invntt_layer7_invert_harness.c
@@ -1,0 +1,30 @@
+// Copyright (c) 2024 The mlkem-native project authors
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+
+/*
+ * Insert copyright notice
+ */
+
+/**
+ * @file invntt_layer7_invert_harness.c
+ * @brief Implements the proof harness for invntt_layer7_invert function.
+ */
+
+/*
+ * Insert project header files that
+ *   - include the declaration of the function
+ *   - include the types needed to declare function arguments
+ */
+#include <ntt.h>
+void invntt_layer7_invert(int16_t *r);
+
+/**
+ * @brief Starting point for formal analysis
+ *
+ */
+void harness(void)
+{
+  int16_t *a;
+  invntt_layer7_invert(a);
+}