--- a/cbmc/ntt_layer45/ntt_layer45_harness.c
+++ b/cbmc/ntt_layer45/ntt_layer45_harness.c
@@ -1,0 +1,30 @@
+// Copyright (c) 2024 The mlkem-native project authors
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+
+/*
+ * Insert copyright notice
+ */
+
+/**
+ * @file ntt_layer45_harness.c
+ * @brief Implements the proof harness for ntt_layer45 function.
+ */
+
+/*
+ * Insert project header files that
+ *   - include the declaration of the function
+ *   - include the types needed to declare function arguments
+ */
+#include <ntt.h>
+void ntt_layer45(int16_t *r);
+
+/**
+ * @brief Starting point for formal analysis
+ *
+ */
+void harness(void)
+{
+  int16_t *a;
+  ntt_layer45(a);
+}